refactor: types 디렉토리 하위 파일 이름 변경
- types 하위 디렉토리는 [name]Type 으로 정의한다.
