+++ src/pages/index.tsx
@@ -1,19 +1,19 @@
 import clsx from 'clsx';
-import User from '@/types/userInterface';
+import User from '@/types/userType';
 import Image from 'next/image';
 import axios from 'axios';
 import styles from './Home.module.scss';
 import Footer from '@/components/footer/Footer';
 import dynamic from 'next/dynamic';
 import Category from '@/components/category/Category';
-import { Store } from '@/types/redux';
+import { Store } from '@/types/reduxType';
 import { login } from '@/actions/actions';
 import { useQuery } from '@tanstack/react-query';
 import ReactPaginate from 'react-paginate';
 import { useRouter } from 'next/router';
 import const_queryKey from '@/const/queryKey';
 import { getIndicators } from '@/backendApi/fred';
-import { Category_Type } from '@/types/fredInterface';
+import { Category_Type } from '@/types/fredType';
 import { roboto, poppins } from './_app';
 import CategoryWithIsActive from '@/components/categoryWithIsAcitve/CategoryWithIsActive';
 import { useEffect, useState } from 'react';
