+++ src/types/reduxType.ts
@@ -1,4 +1,4 @@
-import User from './userInterface';
+import User from './userType';
 
 export type Store = {
 	user: User;
